refactor: 코드 분할 및 useReducer, useVote 사용
