+++ package.json
@@ -8,8 +8,9 @@
     "@testing-library/user-event": "^12.2.0",
     "axios": "^0.21.1",
     "react": "^17.0.1",
+    "react-cookie": "^4.0.3",
     "react-dom": "^17.0.1",
-    "react-scripts": "4.0.0",
+    "react-scripts": "^4.0.3",
     "styled-components": "^5.2.1",
     "web-vitals": "^0.2.4"
   },
